Refactor case and cache services
