+++ src/main/definitions/api/caseApiResponse.ts
@@ -1,0 +1,26 @@
+import { CaseType, YesOrNo } from '../case';
+import { ClaimantCorrespondence } from '../complexTypes/claimantCorrespondence';
+import { ClaimantIndividual } from '../complexTypes/claimantIndividual';
+import { CaseState } from '../definition';
+
+export interface CaseApiResponse {
+  id: string;
+  jurisdiction?: string;
+  state: CaseState;
+  case_type_id?: string;
+  created_date?: Date;
+  last_modified?: Date;
+  locked_by_user_id?: boolean | null;
+  security_level?: string | null;
+  case_data?: CaseData;
+  security_classification?: string;
+  callback_response_status?: string | null;
+}
+
+export interface CaseData {
+  caseType?: CaseType;
+  caseSource?: string;
+  claimantRepresentedQuestion?: YesOrNo;
+  claimantIndType?: ClaimantIndividual;
+  claimantType?: ClaimantCorrespondence;
+}
