+++ src/main/definitions/case.ts
@@ -1,5 +1,3 @@
-import { ClaimantCorrespondence } from './complexTypes/claimantCorrespondence';
-import { ClaimantIndividual } from './complexTypes/claimantIndividual';
 import { CaseState, ClaimOutcomes, TypesOfClaim } from './definition';
 import { UnknownRecord } from './util-types';
 
@@ -113,26 +111,4 @@ export const enum CaseDataCacheKey {
   CLAIMANT_REPRESENTED = 'claimantRepresentedQuestion',
   CASE_TYPE = 'caseType',
   TYPES_OF_CLAIM = 'typesOfClaim',
-}
-
-export interface CaseApiResponse {
-  id: string;
-  jurisdiction?: string;
-  state: CaseState;
-  case_type_id?: string;
-  created_date?: Date;
-  last_modified?: Date;
-  locked_by_user_id?: boolean | null;
-  security_level?: string | null;
-  case_data?: CaseData;
-  security_classification?: string;
-  callback_response_status?: string | null;
-}
-
-export interface CaseData {
-  caseType?: CaseType;
-  caseSource?: string;
-  claimantRepresentedQuestion?: YesOrNo;
-  claimantIndType?: ClaimantIndividual;
-  claimantType?: ClaimantCorrespondence;
 }
