+++ src/main/definitions/constants.ts
@@ -133,10 +133,7 @@ export const CaseApiErrors = {
 } as const;
 
 export const CcdDataModel = {
-  SINGLE_CASE_ENGLAND: 'ET_EnglandWales',
-  MULTIPLE_CASE_ENGLAND: 'ET_EnglandWales_Multiple',
-  SINGLE_CASE_SCOTLAND: 'ET_Scotland',
   CASE_SOURCE: 'ET1 Online',
-};
+} as const;
 
 export const EXISTING_USER = 'existingUser';
